+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/29 18:27:32 by gvial             #+#    #+#             */
-/*   Updated: 2022/10/12 11:10:14 by mraymond         ###   ########.fr       */
+/*   Updated: 2022/10/12 13:27:27 by mraymond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -143,29 +143,6 @@ char	*space_out_redirections(char *last_line);
 //04_builtins
 char	**add_env_var(char **envp, char **vars);
 
-//05_exec.c
-void	exec(t_ms *ms);
-void	fd_redirection(t_ms *ms);
-void	close_ms_pipe(t_ms *ms);
-void	fd_allocation(t_ms *ms);
-void	child_creation(t_ms *ms);
-void	waiting_n_closefd(t_ms *ms);
-int		child_process_to_index(t_ms *ms, int waitpid_return);
-
-//05_child_exec.c
-void	child_execution(t_ms *ms);
-t_cmd	*cmd_lst_index(t_ms *ms, int cmd_index);
-int		pipe_redirection(t_ms *ms, t_cmd *cmd);
-void	exec_fail(t_ms *ms, t_cmd *cmd);
-void	close_keep_errno(int fd);
-
-//05_redirection.c
-int		redirection_in(t_cmd *cmd);
-int		here_doc(char *str_eof);
-int		open_fd_in(char *filename);
-int		print_open_err(char *filename, int error);
-int		redirection_out(t_cmd *cmd);
-
 //01_valid_line
 int		valid_line(char *line);
 int		print_line_err(int error);
@@ -175,6 +152,33 @@ int		lst_len(t_cmd *head);
 
 //utils2
 void	free_dbl_ptr(void **ptr, int option);
+
+//EXECUTION---------------------------------------------------------------------
+//05_exec.c
+void	exec(t_ms *ms);
+void	fd_allocation(t_ms *ms);
+void	fd_redirection(t_ms *ms);
+void	child_creation(t_ms *ms);
+void	waiting_n_closefd(t_ms *ms);
+
+//05_child_exec.c
+void	child_execution(t_ms *ms);
+void	closing_pipe(t_ms *ms);
+int		pipe_redirection(t_ms *ms, t_cmd *cmd);
+void	exec_fail(t_ms *ms, t_cmd *cmd);
+
+//05_redirection.c
+int		redirection_in(t_cmd *cmd);
+int		here_doc(char *str_eof);
+int		open_fd_in(char *filename);
+int		print_open_err(char *filename, int error);
+int		redirection_out(t_cmd *cmd);
+
+//05_exec_utils.c
+int		child_process_to_index(t_ms *ms, int waitpid_return);
+void	close_ms_pipe(t_ms *ms);
+void	close_keep_errno(int fd);
+t_cmd	*cmd_lst_index(t_ms *ms, int cmd_index);
 void	free_ms(t_ms *ms, int with_lstcmds);
 
 //==============================================================================
